requirements.txt
numpy
pillow
requests
tensorflow
scikit-learn